reimplement reward function taking into account z velocity
We use a sigmoid function to check for the difference between the expected z velocity and the current z velocity based on the quadcopter z position.
We obtained significantly better results than the previous reward function.
The next step will be to take into account other velocities (x,y and angular) in order to stabilise the quadcopter during the ascent.
All other velocities should be close to zero, as we want our quadcopter to take off at a predetermined z position and to stay there as long as it can.